# back/services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "陪玩平台 Gateway 服务 API 文档",
        "title": "陪玩平台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
        "/api/agent/recommend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据用户需求，使用AI智能推荐合适的陪玩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI助手"
                ],
                "summary": "AI推荐陪玩",
                "parameters": [
                    {
                        "description": "AI推荐陪玩请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RecommendCompanionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.RecommendCompanionResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/code/send": {
            "post": {
                "description": "向指定手机号发送短信验证码，用于注册、登录、修改密码等场景",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "发送验证码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SendCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.SendCodeResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/order": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据订单ID获取订单详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单ID",
                        "name": "orderId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetOrderResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "老板向陪玩下单，创建陪玩订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "创建订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.CreateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/follow/mutual": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取与当前用户互相关注的用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注"
                ],
                "summary": "获取互相关注列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetMutualFollowListResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/followers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的粉丝列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注"
                ],
                "summary": "获取我的粉丝列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetMyFollowersListResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/user/following": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的关注列表，支持分页和昵称搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注"
                ],
                "summary": "获取我的关注列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（昵称模糊匹配）",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetMyFollowingListResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/unfollow": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消关注指定用户，需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注"
                ],
                "summary": "取消关注",
                "parameters": [
                    {
                        "description": "取消关注请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UnfollowUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UnfollowUserResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.CompanionRecommendation": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "gameSkill": {
                    "description": "游戏技能",
                    "type": "string"
                },
                "gender": {
                    "description": "性别",
                    "type": "string"
                },
                "pricePerHour": {
                    "description": "每小时价格",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分",
                    "type": "number"
                },
                "similarity": {
                    "description": "相似度分数（0-1）",
                    "type": "number"
                },
                "userId": {
                    "description": "陪玩用户ID",
                    "type": "integer"
                }
            }
        },
...
        "types.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "companionId": {
                    "description": "陪玩ID",
                    "type": "integer"
                },
                "durationHours": {
                    "description": "服务时长（小时）",
                    "type": "integer"
                },
                "gameName": {
                    "description": "游戏名称",
                    "type": "string"
                }
            }
        },
        "types.CreateOrderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.OrderInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.GetMutualFollowListData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.UserFollowInfo"
                    }
                }
            }
        },
        "types.GetMutualFollowListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.GetMutualFollowListData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.GetMyFollowersListData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.UserFollowInfo"
                    }
                }
            }
        },
        "types.GetMyFollowersListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.GetMyFollowersListData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.GetMyFollowingListData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.UserFollowInfo"
                    }
                }
            }
        },
        "types.GetMyFollowingListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.GetMyFollowingListData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.GetOrderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.OrderInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.OrderInfo": {
            "type": "object",
            "properties": {
                "acceptedAt": {
                    "description": "接单时间",
                    "type": "integer"
                },
                "bossId": {
                    "description": "老板ID",
                    "type": "integer"
                },
                "cancelReason": {
                    "description": "取消原因",
                    "type": "string"
                },
                "cancelledAt": {
                    "description": "取消时间",
                    "type": "integer"
                },
                "comment": {
                    "description": "评价内容",
                    "type": "string"
                },
                "companionId": {
                    "description": "陪玩ID",
                    "type": "integer"
                },
                "completedAt": {
                    "description": "完成时间",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "durationHours": {
                    "description": "时长（小时）",
                    "type": "integer"
                },
                "gameName": {
                    "description": "游戏名称",
                    "type": "string"
                },
                "id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "orderNo": {
                    "description": "订单号",
                    "type": "string"
                },
                "paidAt": {
                    "description": "支付时间",
                    "type": "integer"
                },
                "pricePerHour": {
                    "description": "每小时价格（帅币）",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分",
                    "type": "number"
                },
                "startAt": {
                    "description": "开始服务时间",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1=CREATED,2=PAID,3=ACCEPTED,4=IN_SERVICE,5=COMPLETED,6=CANCELLED,7=RATED",
                    "type": "integer"
                },
                "totalAmount": {
                    "description": "订单总价（帅币）",
                    "type": "integer"
                }
            }
        },
...
        "types.RecommendCompanionData": {
            "type": "object",
            "properties": {
                "companions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CompanionRecommendation"
                    }
                },
                "explanation": {
                    "type": "string"
                }
            }
        },
        "types.RecommendCompanionRequest": {
            "type": "object",
            "properties": {
                "userInput": {
                    "description": "用户输入内容",
                    "type": "string"
                }
            }
        },
        "types.RecommendCompanionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.RecommendCompanionData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.SendCodeData": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "types.SendCodeRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "types.SendCodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.SendCodeData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UnfollowUserData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "types.UnfollowUserRequest": {
            "type": "object",
            "properties": {
                "userId": {
                    "description": "要取消关注的用户ID",
                    "type": "integer"
                }
            }
        },
        "types.UnfollowUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UnfollowUserData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UserFollowInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "followedAt": {
                    "description": "关注时间戳",
                    "type": "integer"
                },
                "isMutual": {
                    "description": "是否互相关注",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}